+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/dao/cr/MultivalueTypeDao.java
@@ -31,7 +31,7 @@ public class MultivalueTypeDao extends DaoForEntityWithSurrogateKey<MultivalueTy
 	}
 
 	@Override
-	public void deleteRaw(Long id)
+	public void deleteRaw(Integer id)
 	{
 		throw new UnsupportedOperationException(
 			"Cannot delete rows from table: " + getTableName() + "."
