+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/pages/outputws/EditPropertyPage.java
@@ -28,7 +28,7 @@ public class EditPropertyPage extends FrontendPage
 	private DaoForEntityWithSurrogateKey<AggregationType> aggregationTypeDao;
 	private DaoForEntityWithSurrogateKey<MultivalueType> multivalueTypeDao;
 
-	public EditPropertyPage(final Long propertyId) 
+	public EditPropertyPage(final Integer propertyId) 
 	{
 		super(
 			"Home > Output WS > Aggregation Properties > Edit", 
@@ -47,7 +47,7 @@ public class EditPropertyPage extends FrontendPage
 		addEditPropertyForm(propertyId);
 	}
 	
-	private void addEditPropertyForm(final Long propertyId)
+	private void addEditPropertyForm(final Integer propertyId)
 	{
 		PropertySettings property = propertySettingsDao.load(propertyId);
