+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/pages/pipelines/PipelineDetailPage.java
@@ -35,7 +35,7 @@ public class PipelineDetailPage extends FrontendPage
 	private DaoForEntityWithSurrogateKey<Pipeline> pipelineDao;
 	private DaoForEntityWithSurrogateKey<TransformerInstance> transformerInstanceDao;
 
-	public PipelineDetailPage(final Long pipelineId) 
+	public PipelineDetailPage(final Integer pipelineId) 
 	{
 		super(
 			"Home > Backend > Pipelines > Detail", 
@@ -55,7 +55,7 @@ public class PipelineDetailPage extends FrontendPage
 		addAssignmentSection(pipelineId);
 	}
 
-	private void addPipelineInformationSection(final Long pipelineId)
+	private void addPipelineInformationSection(final Integer pipelineId)
 	{
 		setDefaultModel(createModelForOverview(pipelineDao, pipelineId));
 		
@@ -64,7 +64,7 @@ public class PipelineDetailPage extends FrontendPage
 		add(new Label("isDefault"));
 	}
 
-	private void addAssignmentSection(final Long pipelineId) 
+	private void addAssignmentSection(final Integer pipelineId) 
 	{
 		add(
 			new RedirectWithParamButton(
@@ -77,7 +77,7 @@ public class PipelineDetailPage extends FrontendPage
 		addAssignmentTable(pipelineId);
 	}
 	
-	private void addAssignmentTable(final Long pipelineId) 
+	private void addAssignmentTable(final Integer pipelineId) 
 	{
 		SortableDataProvider<TransformerInstance> data = new DependentSortableDataProvider<TransformerInstance>
 		(
